refactor(admin): update card styles and component name - change to KPICard - rework layout - update styles and remove inline styles
